add TronStoreWithRevoking
