+++ src/main/java/org/tron/core/db/AssetIssueStore.java
@@ -5,10 +5,9 @@ import java.util.stream.Collectors;
 import org.apache.commons.lang3.ArrayUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.tron.common.utils.ByteArray;
 import org.tron.core.capsule.AssetIssueCapsule;
 
-public class AssetIssueStore extends TronDatabase<AssetIssueCapsule> {
+public class AssetIssueStore extends TronStoreWithRevoking<AssetIssueCapsule> {
 
   private static final Logger logger = LoggerFactory.getLogger("AssetIssueStore");
   private static AssetIssueStore instance;
@@ -16,23 +15,6 @@ public class AssetIssueStore extends TronDatabase<AssetIssueCapsule> {
 
   private AssetIssueStore(String dbName) {
     super(dbName);
-  }
-
-  @Override
-  public void put(byte[] key, AssetIssueCapsule item) {
-    logger.info("asset issue is {}, asset issue is {}", key, item);
-
-    byte[] value = dbSource.getData(key);
-    if (ArrayUtils.isNotEmpty(value)) {
-      onModify(key, value);
-    }
-
-    logger.info("name is {} ", ByteArray.toHexString(key));
-    dbSource.putData(key, item.getData());
-
-    if (ArrayUtils.isEmpty(value)) {
-      onCreate(key);
-    }
   }
 
   /**
@@ -49,13 +31,6 @@ public class AssetIssueStore extends TronDatabase<AssetIssueCapsule> {
       }
     }
     return instance;
-  }
-
-  @Override
-  public void delete(byte[] key) {
-    // This should be called just before an object is removed.
-    onDelete(key);
-    dbSource.deleteData(key);
   }
 
   @Override
