+++ src/main/java/org/tron/core/db/BlockStore.java
@@ -20,11 +20,10 @@ import org.joda.time.DateTime;
 import org.joda.time.format.DateTimeFormat;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.tron.common.utils.ByteArray;
 import org.tron.common.utils.Sha256Hash;
 import org.tron.core.capsule.BlockCapsule;
 
-public class BlockStore extends TronDatabase<BlockCapsule> {
+public class BlockStore extends TronStoreWithRevoking<BlockCapsule> {
 
   public static final Logger logger = LoggerFactory.getLogger("BlockStore");
 
@@ -79,30 +78,6 @@ public class BlockStore extends TronDatabase<BlockCapsule> {
   // genesis_time
   public DateTime getGenesisTime() {
     return DateTime.parse("20180101", DateTimeFormat.forPattern("yyyyMMdd"));
-  }
-
-  @Override
-  public void put(byte[] key, BlockCapsule item) {
-    logger.info("address is {},account is {}", key, item);
-
-    byte[] value = dbSource.getData(key);
-    if (ArrayUtils.isNotEmpty(value)) {
-      onModify(key, value);
-    }
-
-    logger.info("address is {} ", ByteArray.toHexString(key));
-    dbSource.putData(key, item.getData());
-
-    if (ArrayUtils.isEmpty(value)) {
-      onCreate(key);
-    }
-  }
-
-  @Override
-  public void delete(byte[] key) {
-    // This should be called just before an object is removed.
-    onDelete(key);
-    dbSource.deleteData(key);
   }
 
   @Override
