+++ src/main/java/org/tron/core/db/WitnessStore.java
@@ -7,35 +7,12 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.tron.core.capsule.WitnessCapsule;
 
-public class WitnessStore extends TronDatabase<WitnessCapsule> {
+public class WitnessStore extends TronStoreWithRevoking<WitnessCapsule> {
 
   private static final Logger logger = LoggerFactory.getLogger("WitnessStore");
 
   protected WitnessStore(String dbName) {
     super(dbName);
-  }
-
-  @Override
-  public void put(byte[] key, WitnessCapsule item) {
-    logger.info("voteAddress is {} ", item.getAddress());
-
-    byte[] value = dbSource.getData(key);
-    if (ArrayUtils.isNotEmpty(value)) {
-      onModify(key, value);
-    }
-
-    dbSource.putData(key, item.getData());
-
-    if (ArrayUtils.isEmpty(value)) {
-      onCreate(key);
-    }
-  }
-
-  @Override
-  public void delete(byte[] key) {
-    // This should be called just before an object is removed.
-    onDelete(key);
-    dbSource.deleteData(key);
   }
 
   @Override
